Clean: Simplify scrolling - remove smooth animation, int scrollOffset, reduce redundant calculations

+++ OrangeX/src/main/java/cn/xiaozhou233/orangex/ui/clickgui/component/Component.java
@@ -72,7 +72,7 @@ public abstract class Component {
 
 
     public int getY() {
-        return parent.getY() + offsetY - (int) parent.getScrollOffset();
+        return parent.getY() + offsetY - parent.getScrollOffset();
     }
